Tweaked timings further. Setting timings to cycle times of 25ms instead of 10ms. This in the end causes 10ms yields to the CPU which keeps the CPU at the expected levels. These changes result from switching to PrecisionDelay instead of using ThreadDelay. The changes are aimed at having low CPU during times of inactivity and yet allow for spiking to low-latency cycles when things get busy. The main mechanism through which code should accomplish this is by RequestPulse() which bypasses any thread yields.

+++ platform/gucefMT/src/gucefMT_dvmtoswrap.c
@@ -351,7 +351,7 @@ PrecisionDelay( UInt32 delayInMs )
         if ( ticksLeft > (Int32) ( highPerfTimerFrequencyTicksPerMs * 50 ) )
         {
             // Lets not hog the CPU if the time we have to wait is lengthy
-            UInt32 longSleepTimeInMs = (ticksLeft/highPerfTimerFrequencyTicksPerMs) - (highPerfTimerFrequencyTicksPerMs*15);
+            UInt32 longSleepTimeInMs = (UInt32) ( (ticksLeft/highPerfTimerFrequencyTicksPerMs) - (highPerfTimerFrequencyTicksPerMs*15) );
             Sleep( longSleepTimeInMs );
         }
         else
